break out scripts and styles tasks into separate configs
