+++ gulpfile.js
@@ -1,5 +1,7 @@
-// README: Split this file up if it starts getting too long. ( i.e. 300 lines )
-//
+// Bring in individual gulp configs
+
+require('./config/gulp/styles');
+require('./config/gulp/scripts');
 
 // This gulp asset pipeline supports linting of scss and js files along with
 // compiling and bundling css and js files into static/assets/ directories to
@@ -23,7 +25,7 @@ var sassFiles = filter([ '**/*.scss' ], { restore: true });
 var jsFiles = filter([ '**/*.js' ], { restore: true });
 var spawn = require('child_process').spawn;
 
-var cFlags = {
+global.cFlags = {
   production: false,
   test: true,
 };
@@ -47,20 +49,6 @@ gulp.task('clean-all', function () {
   ]);
 });
 
-//gulp.task('scss-lint', function (done) {
-
-  //if (!cFlags.test) {
-    //gutil.log(gutil.colors.cyan('scss-lint'), 'Disabling linting');
-    //return done();
-  //}
-
-  //return gulp.src('./assets/styles/**/*.scss')
-    //.pipe(scsslint({
-      //config: './.scss-lint.yml',
-    //}));
-
-//});
-
 //gulp.task('eslint', function (done) {
 
   //if (!cFlags.test) {
@@ -75,40 +63,6 @@ gulp.task('clean-all', function () {
     //.pipe(eslint.format());
 
 //});
-
-gulp.task('styles:homepage', [ /*'scss-lint'*/ ], function () {
-
-  gutil.log(gutil.colors.cyan('styles:homepage'), 'Compiling Sass assets');
-
-  var sassStream = sass();
-  var stream = gulp.src('./assets/styles/main.scss');
-
-  if (cFlags.production) {
-    gutil.log(gutil.colors.cyan('styles:homepage'), 'Compressing styles');
-    sassStream = sass({ outputStyle: 'compressed' });
-  }
-
-  stream = stream.pipe(sassStream)
-    .on('error', function (error) {
-      gutil.log(
-        gutil.colors.yellow('styles:homepage'),
-        gutil.colors.red('error'),
-        '\n',
-        error.messageFormatted
-      );
-
-      if (cFlags.production) {
-        process.exit(1);
-      }
-
-      this.emit('end');
-    })
-    .pipe(gulp.dest('./static/assets/styles'));
-
-  return stream;
-
-
-});
 
 //gulp.task('scripts', [ 'eslint' ], function () {
